requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
PyYAML==6.0.2

tqdm==4.66.6
faiss-cpu==1.9.0
pypdf==5.1.0
numpy==1.26.4

openai==1.57.1
tiktoken==0.8.0
unstructured[md]==0.16.11
nltk==3.9.1

langchain==0.3.11
langchain-community==0.3.11
langchain-openai==0.2.12
# langchain-mistralai==0.2.3
langchain-chroma==0.1.4
